refactor: extract reusable contact form
