+++ src/routes/contact.tsx
@@ -1,20 +1,6 @@
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
+import { ContactForm } from "@/components/contact-form";
 import { createFileRoute } from "@tanstack/react-router";
 import { useTranslate } from "@tolgee/react";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import { toast } from "sonner";
 
 export const Route = createFileRoute("/contact")({
   component: ContactPage,
@@ -22,29 +8,6 @@ export const Route = createFileRoute("/contact")({
 
 function ContactPage() {
   const { t } = useTranslate();
-
-  const formSchema = z.object({
-    email: z
-      .string()
-      .nonempty({ message: t("validation.required") })
-      .email({ message: t("validation.invalidEmail") }),
-    message: z.string().nonempty({ message: t("validation.required") }),
-  });
-
-  type FormValues = z.infer<typeof formSchema>;
-
-  const formMethods = useForm<FormValues>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      email: "",
-      message: "",
-    },
-  });
-
-  const onSubmit = () => {
-    toast.success(t("contact.success"));
-    formMethods.reset();
-  };
 
   return (
     <div className="flex min-h-dvh flex-col items-center justify-center bg-muted px-4 py-10">
@@ -57,50 +20,7 @@ function ContactPage() {
           />
           <h1 className="text-3xl font-bold">{t("contact.title")}</h1>
           <p className="text-muted-foreground">{t("contact.description")}</p>
-          <Form {...formMethods}>
-            <form
-              onSubmit={formMethods.handleSubmit(onSubmit)}
-              className="flex w-full flex-col gap-4"
-            >
-              <FormField
-                control={formMethods.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>{t("contact.email")}</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="email"
-                        placeholder={t("contact.emailPlaceholder")}
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={formMethods.control}
-                name="message"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>{t("contact.message")}</FormLabel>
-                    <FormControl>
-                      <Textarea
-                        placeholder={t("contact.messagePlaceholder")}
-                        className="min-h-[120px]"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <Button type="submit" className="w-full">
-                {t("contact.submit")}
-              </Button>
-            </form>
-          </Form>
+          <ContactForm className="w-full" />
         </div>
       </div>
     </div>
